chore: use transactionstatus enum for objects
