+++ backend/protocol_rpc/server.py
@@ -43,9 +43,7 @@ def create_app():
     socketio = SocketIO(app, cors_allowed_origins="*")
     msg_handler = MessageHandler(app, socketio)
     genlayer_db_client = DBClient(database_name_seed)
-    transactions_processor = TransactionsProcessor(
-        sqlalchemy_db.session, genlayer_db_client
-    )
+    transactions_processor = TransactionsProcessor(sqlalchemy_db.session)
     accounts_manager = AccountsManager(sqlalchemy_db.session, transactions_processor)
     validators_registry = ValidatorsRegistry(sqlalchemy_db.session)
